+++ database/migrations/tenant/2021_10_15_171401_add_necessary_fields_to_product_variations_table.php
@@ -14,10 +14,16 @@ class AddNecessaryFieldsToProductVariationsTable extends Migration
     public function up()
     {
         Schema::table('product_variations', function (Blueprint $table) {
-            $table->integer('num_of_sales')->after('price')->default(0);
-            $table->timestamp('deleted_at', 0)->nullable(true)->after('updated_at');
-
-            $table->unique(['product_id', 'variant']);
+            if (Schema::hasColumn('product_variations', 'num_of_sales')) {
+                $table->dropColumn('num_of_sales');
+            }
+            if (!Schema::hasColumn('product_variations', 'deleted_at')) {
+                $table->timestamp('deleted_at', 0)->nullable(true)->after('updated_at');
+            }
+            // Add lenght to text field because , text field with no lenght can't be unique key index
+            $table->text('variant', 65535)->change();
+            /* TODO: Fix this, removed it beause text field has an issue for creating this index */
+            // $table->unique(['product_id', 'variant']);
         });
     }
 
@@ -29,10 +35,15 @@ class AddNecessaryFieldsToProductVariationsTable extends Migration
     public function down()
     {
         Schema::table('product_variations', function (Blueprint $table) {
-            $table->dropColumn('num_of_sales');
-            $table->dropColumn('deleted_at');
+            if (Schema::hasColumn('product_variations', 'num_of_sales')) {
+                $table->dropColumn('num_of_sales');
+            }
 
-            $table->dropUnique('product_variations_product_id_variant_index');
+            if (Schema::hasColumn('product_variations', 'deleted_at')) {
+                $table->dropColumn('deleted_at');
+            }
+
+            // $table->dropUnique('product_variations_product_id_variant_index');
         });
     }
 }
